Migrate fjs/crypto/sign to .f.mjs, split types into types.ts
Splits the All type into a sibling types.ts, and converts the runtime
implementation (all, fromCurve, concat, computeK, sign) to JSDoc-typed
module.f.mjs/proof.f.mjs under the stage-1 TypeScript-to-mjs migration.
No other module imports fjs/crypto/sign/module.f.ts; fixes three stale
doc references in fjs/crypto/sign/todo and fjs/crypto/todo.
