+++ fjs/crypto/sign/types.ts
@@ -1,0 +1,15 @@
+/**
+ * Type-level API for signing helpers built on secp256k1 and SHA-256 primitives.
+ *
+ * @module
+ */
+
+import type { Vec } from '../../types/bit_vec/types.ts'
+
+export type All = {
+    readonly q: bigint
+    readonly qlen: bigint
+    readonly bits2int: (b: Vec) => bigint
+    readonly int2octets: (x: bigint) => Vec
+    readonly bits2octets: (b: Vec) => Vec
+}
